requirements.txt
Django==3.2
djangorestframework==3.12.4
python-dotenv==0.17.0
psycopg2==2.8.6
